+++ example/lib/tag.dart
@@ -1,5 +1,8 @@
 class Tag {
-  static const ford = Tag(pubId: "609a943be65f6b2a0c3ffbe5", tagId: "63fc78e436e14ce9ad0f5a66");
+  static const ford = Tag(
+    pubId: "609a943be65f6b2a0c3ffbe5",
+    tagId: "63fc78e436e14ce9ad0f5a66",
+  );
 
   final String pubId;
   final String tagId;