+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 ## 0.1.0
 
-* TODO: Describe initial release.
+- added in-read ads support
+- added play ad feature
+- added pause ad feature
+- added display ad fullscreen feature
+- added ad state listener
+- added ad events listener
+
+## 0.2.0
+
+- added interstitial ads support
+- improved documentation
+